+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Breaking changes:
 New features and notable changes:
 
 - In Azure pipelines or GitHub actions errors and warnings are printed in an additional format captured by the CI. (:issue:`904`)
+- Detect suspicious counter values in ``gcov`` output. (:issue:`903`)
 
 Bug fixes and small improvements: